Урок 4. Макеты Удалили лишние файлы. Переделываем под новые задачи.
