+++ app/src/main/java/by/ealipatov/kotlin/materialyoufromealipatov/MainActivity.kt
@@ -12,8 +12,8 @@ import androidx.fragment.app.Fragment
 import by.ealipatov.kotlin.materialyoufromealipatov.databinding.ActivityMainBinding
 import by.ealipatov.kotlin.materialyoufromealipatov.utils.*
 import by.ealipatov.kotlin.materialyoufromealipatov.view.SettingFragment
-import by.ealipatov.kotlin.materialyoufromealipatov.view.ViewPager.ViewPagerFragment
-import by.ealipatov.kotlin.materialyoufromealipatov.view.ViewPagerPOD.PictureOfTheDayViewPagerFragment
+import by.ealipatov.kotlin.materialyoufromealipatov.view.viewPager.ViewPagerFragment
+import by.ealipatov.kotlin.materialyoufromealipatov.view.viewPagerPOD.PictureOfTheDayViewPagerFragment
 
 class MainActivity : AppCompatActivity() {
