+++ app/src/main/java/by/ealipatov/kotlin/materialyoufromealipatov/view/viewPager/LayoutsFragment.kt
@@ -1,0 +1,34 @@
+package by.ealipatov.kotlin.materialyoufromealipatov.view.viewPager
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import by.ealipatov.kotlin.materialyoufromealipatov.databinding.FragmentLayoutsBinding
+
+class LayoutsFragment: Fragment() {
+
+    private var _binding: FragmentLayoutsBinding? = null
+    private val binding: FragmentLayoutsBinding
+        get() {
+            return _binding!!
+        }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentLayoutsBinding.inflate(inflater)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+
+
+
+}
