+++ app/src/main/java/by/ealipatov/kotlin/materialyoufromealipatov/view/viewPager/ViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package by.ealipatov.kotlin.materialyoufromealipatov.view.ViewPager
+package by.ealipatov.kotlin.materialyoufromealipatov.view.viewPager
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
