--- app/src/main/res/menu/menu_bottom_bar_view_pager_fragment.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/app_bar_back"
-        android:icon="@drawable/ic_back_fab"
-        android:title="@string/back"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/app_bar_settings"
-        android:icon="@drawable/ic_settings_menu"
-        android:title="@string/settings"
-        app:showAsAction="ifRoom" />
-</menu>
